+++ cmd/brabble/main.go
@@ -33,6 +33,7 @@ func run() error {
 	root.AddCommand(control.NewListMicsCmd())
 	root.AddCommand(control.NewSetMicCmd(cfgPath))
 	root.AddCommand(control.NewTestHookCmd(cfgPath))
+	root.AddCommand(control.NewDoctorCmd(cfgPath))
 
 	// Hidden internal serve command used by start.
 	root.AddCommand(daemon.NewServeCmd(cfgPath))